moved signing code to mixin
